Experiment Conclusion
Turns out, the change that I made to the main branch did not show up in Branch#2, despite it Branch#2 being branched off from it. The commit history was not changed in Branch#2. This means that the current history of commit of a super branch will show up in the sub-branch. However, any changes done to the super branch afterward will not appear in the sub branch, and the sub branch's commit history will not change.